change status added

+++ controllers/feed.js
@@ -4,7 +4,8 @@ const path = require('path');
 
 const { validationResult } = require('express-validator');
 
-const Post = require('../models/post')
+const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -52,20 +53,28 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path.replace('\\', '/');
   const title = req.body.title;
   const content = req.body.content;
-  console.log(title, content);
+  let creator;
 
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'Yuriy' },
+    creator: req.userId,
   });
   post.save()
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post)
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator: {_id: creator._id, name: creator.name}
       });
     })
     .catch(err => {
@@ -122,6 +131,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!')
+        error.statusCode = 403;
+        throw error
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -151,12 +165,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      //Check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!')
+        error.statusCode = 403;
+        throw error
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
+       return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Post delete" })
     })
     .catch(err => {
@@ -166,6 +190,7 @@ exports.deletePost = (req, res, next) => {
       next(err);
     })
 };
+
 
 const clearImage = filePath => {
   filePath = path.join(__dirname, '..', filePath);
